+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- [#847]: `decoder`: Fix log format width specifier not working as expected
+
+[#847]: https://github.com/knurling-rs/defmt/pull/847
+
 ## [v0.3.8] - 2024-05-17
 
 - [#840]: `defmt`: Support pre-1.77